refactor(store): split methods across topic files (objects / status / artifacts / events / rendered)
Iter-13 senior-eng review flagged pkg/store/store.go as a 467-LOC
god-file mixing five concerns. The Store type itself stays — every
key is a NamedResource and one struct is the right substrate — but
the methods now live next to their topic.

Same package, same API, same external behavior. Pure file moves
plus the imports cleanup that follows.

  store.go     467 → 212 LOC  (type + New + nameKey + object CRUD)
  status.go    108 → 202 LOC  (+ UpdateStatus / SetCondition / Get* /
                                FailedResources / conditionEqual)
  artifact.go   60 →  89 LOC  (+ SetArtifact / GetArtifact)
  events.go     58 → 187 LOC  (+ AddListener / replayInto / fire /
                                safeInvoke / listenerSet implementation)
  rendered.go    new 26 LOC   (MarkRendered / WasRendered — orphan
                                bookkeeping kept separate from the
                                object surface so its purpose stays
                                discoverable)

Each topic file is now navigable in isolation. Future type-level
carving (Objects / Status / Bus boundaries) is easier with the
concerns physically separated; this PR doesn't take that step
because it would break every consumer's import path and the file-
level split already captures the readability win.

`go test -race ./pkg/store/... ./pkg/orchestrator/...
./pkg/controllers/...` clean under -race.
