+++ pkg/store/store.go
@@ -1,9 +1,7 @@
 package store
 
 import (
-	"maps"
 	"reflect"
-	"slices"
 	"sync"
 
 	"github.com/home-operations/flate/pkg/manifest"
@@ -74,24 +72,6 @@ func New() *Store {
 			EventArtifactUpdated: newListenerSet(),
 		},
 	}
-}
-
-// MarkRendered records that id was emitted by a parent Kustomization's
-// render output (as opposed to coming only from the static file walk).
-// The orchestrator's post-run phase uses this to distinguish real
-// failures from orphaned-and-loaded-anyway resources.
-func (s *Store) MarkRendered(id manifest.NamedResource) {
-	s.mu.Lock()
-	s.rendered[id] = struct{}{}
-	s.mu.Unlock()
-}
-
-// WasRendered reports whether id was ever emitted by a parent's render.
-func (s *Store) WasRendered(id manifest.NamedResource) bool {
-	s.mu.RLock()
-	_, ok := s.rendered[id]
-	s.mu.RUnlock()
-	return ok
 }
 
 func nameKey(namespace, name string) string { return namespace + "/" + name }
@@ -230,238 +210,3 @@ func (s *Store) ListObjects(kind string) []manifest.BaseManifest {
 	return out
 }
 
-// UpdateStatus records a Ready-condition transition and dispatches a
-// StatusUpdated event when the StatusInfo rollup changes. Internally
-// the (status, message) pair is stored as a metav1.Condition so
-// future callers (ReadyExpr CEL, SOPS detection, healthChecks) can
-// see the rich state via GetConditions.
-func (s *Store) UpdateStatus(id manifest.NamedResource, status Status, message string) {
-	s.SetCondition(id, readyCondition(status, message))
-}
-
-// SetCondition upserts cond into the resource's condition list keyed
-// by cond.Type. Dispatches a StatusUpdated event with the *StatusInfo
-// rollup* (derived from Ready) on every observable condition change,
-// not just Ready transitions. Listeners that only care about Ready
-// can filter on the StatusInfo payload; CEL-based ReadyExpr watchers
-// need the broader notification so a Healthy condition flip (for
-// example) wakes them.
-//
-// An identical re-write of the same condition is a no-op (no event).
-func (s *Store) SetCondition(id manifest.NamedResource, cond Condition) {
-	s.mu.Lock()
-	prev := s.conditions[id]
-
-	updated := make([]Condition, 0, len(prev)+1)
-	replaced := false
-	for _, c := range prev {
-		if c.Type == cond.Type {
-			if conditionEqual(c, cond) {
-				// Identical condition — nothing to do, including no event.
-				s.mu.Unlock()
-				return
-			}
-			updated = append(updated, cond)
-			replaced = true
-			continue
-		}
-		updated = append(updated, c)
-	}
-	if !replaced {
-		updated = append(updated, cond)
-	}
-	s.conditions[id] = updated
-	newInfo, _ := statusInfoFromConditions(updated)
-	s.mu.Unlock()
-
-	s.fire(EventStatusUpdated, id, newInfo)
-}
-
-// GetStatus returns the Ready-derived StatusInfo for id and whether
-// a Ready condition was present.
-func (s *Store) GetStatus(id manifest.NamedResource) (StatusInfo, bool) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return statusInfoFromConditions(s.conditions[id])
-}
-
-// GetConditions returns a copy of id's condition list. Empty for
-// unknown ids.
-func (s *Store) GetConditions(id manifest.NamedResource) []Condition {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	conds := s.conditions[id]
-	if len(conds) == 0 {
-		return nil
-	}
-	out := make([]Condition, len(conds))
-	copy(out, conds)
-	return out
-}
-
-// conditionEqual reports whether two conditions carry the same
-// observable state. LastTransitionTime is intentionally ignored — it
-// is reset on every transition by the controller-runtime libraries
-// and would otherwise prevent the no-op short-circuit from firing.
-func conditionEqual(a, b Condition) bool {
-	return a.Type == b.Type &&
-		a.Status == b.Status &&
-		a.Reason == b.Reason &&
-		a.Message == b.Message &&
-		a.ObservedGeneration == b.ObservedGeneration
-}
-
-// SetArtifact stores an artifact for id and dispatches an
-// ArtifactUpdated event. Re-setting with a deep-equal value is a no-op.
-func (s *Store) SetArtifact(id manifest.NamedResource, artifact Artifact) {
-	s.mu.Lock()
-	prev, exists := s.artifacts[id]
-	if exists && reflect.DeepEqual(prev, artifact) {
-		s.mu.Unlock()
-		return
-	}
-	s.artifacts[id] = artifact
-	s.mu.Unlock()
-	s.fire(EventArtifactUpdated, id, artifact)
-}
-
-// GetArtifact returns the artifact for id, or nil if none was set.
-func (s *Store) GetArtifact(id manifest.NamedResource) Artifact {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.artifacts[id]
-}
-
-// FailedResources returns every (id, info) currently in Failed state.
-func (s *Store) FailedResources() map[manifest.NamedResource]StatusInfo {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	out := make(map[manifest.NamedResource]StatusInfo)
-	for id, conds := range s.conditions {
-		if info, ok := statusInfoFromConditions(conds); ok && info.Status == StatusFailed {
-			out[id] = info
-		}
-	}
-	return out
-}
-
-// AddListener registers a callback for the given event kind. When
-// flush==true, the listener is immediately invoked with every matching
-// object already in the store before the call returns. Replay order is
-// unspecified (Go-map iteration); listeners that need a deterministic
-// order must sort what they receive. Listener panics during replay
-// are recovered, same as live dispatch. The returned Unsubscribe
-// removes the listener.
-func (s *Store) AddListener(event EventKind, fn Listener, flush bool) Unsubscribe {
-	set, ok := s.listeners[event]
-	if !ok {
-		panic("store: unknown event kind")
-	}
-	handle := set.add(fn)
-
-	if flush {
-		s.replayInto(event, fn)
-	}
-	return func() { set.remove(handle) }
-}
-
-// replayInto delivers existing state to a fresh listener so it can catch
-// up. Called under no lock; takes its own RLock. Listener panics are
-// recovered via safeInvoke — replay must NOT crash the controller that
-// just attached the listener (parity with live `fire` dispatch).
-func (s *Store) replayInto(event EventKind, fn Listener) {
-	switch event {
-	case EventObjectAdded:
-		s.mu.RLock()
-		snap := make(map[manifest.NamedResource]manifest.BaseManifest, len(s.objects))
-		maps.Copy(snap, s.objects)
-		s.mu.RUnlock()
-		for id, obj := range snap {
-			safeInvoke(fn, id, obj)
-		}
-	case EventStatusUpdated:
-		s.mu.RLock()
-		snap := make(map[manifest.NamedResource]StatusInfo, len(s.conditions))
-		for id, conds := range s.conditions {
-			if info, ok := statusInfoFromConditions(conds); ok {
-				snap[id] = info
-			}
-		}
-		s.mu.RUnlock()
-		for id, info := range snap {
-			safeInvoke(fn, id, info)
-		}
-	case EventArtifactUpdated:
-		s.mu.RLock()
-		snap := make(map[manifest.NamedResource]Artifact, len(s.artifacts))
-		maps.Copy(snap, s.artifacts)
-		s.mu.RUnlock()
-		for id, art := range snap {
-			safeInvoke(fn, id, art)
-		}
-	}
-}
-
-// fire dispatches an event to every registered listener. Listeners run
-// synchronously on the calling goroutine. A panic in any listener is
-// recovered to prevent one bad listener from breaking the dispatch.
-func (s *Store) fire(event EventKind, id manifest.NamedResource, payload any) {
-	set := s.listeners[event]
-	if set == nil {
-		return
-	}
-	for _, fn := range set.snapshot() {
-		safeInvoke(fn, id, payload)
-	}
-}
-
-func safeInvoke(fn Listener, id manifest.NamedResource, payload any) {
-	defer func() { _ = recover() }()
-	fn(id, payload)
-}
-
-// listenerSet is a copy-on-snapshot slice of listeners. add returns a
-// handle (a stable id) used by remove to find the entry. We deliberately
-// do not reuse handles after removal to avoid ABA bugs in long sessions.
-type listenerSet struct {
-	mu      sync.Mutex
-	entries []listenerEntry
-	nextID  int64
-}
-
-type listenerEntry struct {
-	id int64
-	fn Listener
-}
-
-func newListenerSet() *listenerSet { return &listenerSet{} }
-
-func (l *listenerSet) add(fn Listener) int64 {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	l.nextID++
-	id := l.nextID
-	l.entries = append(l.entries, listenerEntry{id: id, fn: fn})
-	return id
-}
-
-func (l *listenerSet) remove(id int64) {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	l.entries = slices.DeleteFunc(l.entries, func(e listenerEntry) bool {
-		return e.id == id
-	})
-}
-
-// snapshot returns a copy of the current listener funcs so dispatch can
-// iterate without holding the lock (and so listeners can mutate the set
-// during dispatch without affecting the current pass).
-func (l *listenerSet) snapshot() []Listener {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	out := make([]Listener, len(l.entries))
-	for i := range l.entries {
-		out[i] = l.entries[i].fn
-	}
-	return out
-}
